+++ mongodb-datastore/common/errors.go
@@ -1,4 +1,4 @@
-package db
+package common
 
 import "fmt"
