+++ mongodb-datastore/handlers/events.go
@@ -1,70 +1,11 @@
 package handlers
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"github.com/keptn/keptn/mongodb-datastore/db"
-	logger "github.com/sirupsen/logrus"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/jeremywohl/flatten"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
+	"github.com/keptn/keptn/mongodb-datastore/db"
 	"github.com/keptn/keptn/mongodb-datastore/models"
 	"github.com/keptn/keptn/mongodb-datastore/restapi/operations/event"
 )
-
-const (
-	contextToProjectCollection        = "contextToProject"
-	rootEventCollectionSuffix         = "-rootEvents"
-	invalidatedEventsCollectionSuffix = "-invalidatedEvents"
-	unmappedEventsCollectionName      = "keptnUnmappedEvents"
-	keptn07EvaluationDoneEventType    = "sh.keptn.events.evaluation-done"
-)
-
-var (
-	client *mongo.Client
-	mutex  sync.Mutex
-
-	projectLocks = map[string]*sync.Mutex{}
-	// define the indexes that should be created for each collection
-	rootEventsIndexes        = []string{"data.service", "time"}
-	projectEventsIndexes     = []string{"data.service", "shkeptncontext", "type"}
-	invalidatedEventsIndexes = []string{"triggeredid"}
-	// keep track of created indexes in memory to save some calls to the mongodb API
-	skipCreateIndex = map[string]bool{}
-	mongoDBName     string
-)
-
-// LockProject locks the collections for a project
-func LockProject(project string) {
-	if projectLocks[project] == nil {
-		mutex.Lock()
-		defer mutex.Unlock()
-		projectLocks[project] = &sync.Mutex{}
-	}
-
-	projectLocks[project].Lock()
-}
-
-// UnLockProject unlocks the collections for a project
-func UnlockProject(project string) {
-	if projectLocks[project] == nil {
-		mutex.Lock()
-		defer mutex.Unlock()
-		projectLocks[project] = &sync.Mutex{}
-	}
-
-	projectLocks[project].Unlock()
-}
 
 type ProjectEventData struct {
 	Project *string `json:"project,omitempty"`
@@ -78,9 +19,9 @@ func NewEventRequestHandler(eventRepo db.EventRepo) *EventRequestHandler {
 	return &EventRequestHandler{eventRepo: eventRepo}
 }
 
-func (erh *EventRequestHandler) SaveEvent(event *models.KeptnContextExtendedCE) error {
+func (erh *EventRequestHandler) ProcessEvent(event *models.KeptnContextExtendedCE) error {
 	if string(event.Type) == keptnv2.GetFinishedEventType(keptnv2.ProjectDeleteTaskName) {
-		return erh.eventRepo.DropProjectCollections(getProjectOfEvent(event))
+		return erh.eventRepo.DropProjectCollections(*event)
 	}
 
 	return erh.eventRepo.InsertEvent(*event)
@@ -94,832 +35,10 @@ func (erh *EventRequestHandler) GetEvents(params event.GetEventsParams) (*event.
 	return (*event.GetEventsOKBody)(events), nil
 }
 
-//==================================
-
-func ensureDBConnection() error {
-	mutex.Lock()
-	defer mutex.Unlock()
-	var err error
-	if client == nil {
-		logger.Debug("No MongoDB client has been initialized yet. Creating a new one.")
-		return connectMongoDBClient()
-	} else if err = client.Ping(context.TODO(), nil); err != nil {
-		logger.Debug("MongoDB client lost connection. Attempt reconnect.")
-		return connectMongoDBClient()
-	}
-
-	return nil
-}
-
-func connectMongoDBClient() error {
-	var err error
-	connectionString, dbName, err := GetMongoDBConnectionString()
-	if err != nil {
-		return fmt.Errorf("failed to create mongo client: %v", err)
-	}
-	mongoDBName = dbName
-	client, err = mongo.NewClient(options.Client().ApplyURI(connectionString))
-	if err != nil {
-		err := fmt.Errorf("failed to create mongo client: %v", err)
-		return err
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-
-	defer cancel()
-
-	err = client.Connect(ctx)
-	if err != nil {
-		err := fmt.Errorf("failed to connect client to MongoDB: %v", err)
-		return err
-	}
-
-	return nil
-}
-
-// ProcessEvent processes the passed event.
-func ProcessEvent(event *models.KeptnContextExtendedCE) error {
-	logger.Debug("save event to data store")
-
-	if err := ensureDBConnection(); err != nil {
-		err := fmt.Errorf("failed to establish MongoDB connection: %v", err)
-		logger.Error(err.Error())
-		return err
-	}
-
-	if string(event.Type) == keptnv2.GetFinishedEventType(keptnv2.ProjectDeleteTaskName) {
-		return dropProjectEvents(event)
-	}
-
-	return insertEvent(event)
-}
-
-func insertEvent(event *models.KeptnContextExtendedCE) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	// insert in project collection
-	collectionName := getProjectOfEvent(event)
-
-	LockProject(collectionName)
-	defer UnlockProject(collectionName)
-
-	collection := client.Database(mongoDBName).Collection(collectionName)
-
-	logger.Debug("Storing event to collection " + collectionName)
-
-	eventInterface, err := transformEventToInterface(event)
-	if err != nil {
-		err := fmt.Errorf("failed to transform event: %v", err)
-		logger.Error(err.Error())
-		return err
-	}
-
-	// additionally store "invalidated" events in a dedicated collection
-	if strings.HasSuffix(string(event.Type), ".invalidated") {
-		invalidatedCollectionName := getInvalidatedCollectionName(collectionName)
-		logger.Debug("Storing invalidated event to dedicated collection " + invalidatedCollectionName)
-		invalidatedCollection := client.Database(mongoDBName).Collection(invalidatedCollectionName)
-
-		ensureIndexExistsOnCollection(
-			ctx,
-			invalidatedCollection,
-			"triggeredid",
-		)
-		_, err = invalidatedCollection.InsertOne(ctx, eventInterface)
-		if err != nil {
-			err := fmt.Errorf("failed to insert into collection: %v", err)
-			logger.Error(err.Error())
-			return err
-		}
-	}
-
-	for _, indexName := range projectEventsIndexes {
-		ensureIndexExistsOnCollection(
-			ctx,
-			collection,
-			indexName,
-		)
-	}
-
-	res, err := collection.InsertOne(ctx, eventInterface)
-	if err != nil {
-		err := fmt.Errorf("failed to insert into collection: %v", err)
-		logger.Error(err.Error())
-		return err
-	}
-	logger.Debug(fmt.Sprintf("insertedID: %s", res.InsertedID))
-
-	err = storeContextToProjectMapping(event, ctx, collectionName)
-	if err != nil {
-		return err
-	}
-
-	err = storeRootEvent(collectionName, ctx, event)
-	if err != nil {
-		return err
-	}
-
-	logger.Debug(fmt.Sprintf("inserted mapping %s->%s", event.Shkeptncontext, collectionName))
-	return nil
-}
-
-func ensureIndexExistsOnCollection(ctx context.Context, collection *mongo.Collection, indexName string) {
-	logger.Debug("ensuring index for " + collection.Name() + " exists")
-	indexID := getIndexIDForCollection(collection.Name(), indexName)
-
-	// if this index has already been created, there is no need to do so again
-	if skipCreateIndex[indexID] {
-		return
-	}
-
-	indexDefinition := mongo.IndexModel{
-		Keys: bson.M{
-			indexName: 1,
-		},
-	}
-	// CreateOne() is idempotent - this operation checks if the index exists and only creates a new one when not available
-	createdIndex, err := collection.Indexes().CreateOne(ctx, indexDefinition)
-	if err != nil {
-		// log the error, but continue anyway - index is not required for the query to work
-		logger.Debug("could not create index for " + collection.Name() + ": " + err.Error())
-	}
-	fmt.Println(createdIndex)
-	// keep track (in memory) that this index already exists
-	skipCreateIndex[indexID] = true
-	logger.Debug("created index for " + collection.Name())
-}
-
-func getIndexIDForCollection(collectionName string, indexName string) string {
-	return collectionName + "-" + indexName
-}
-
-func getInvalidatedCollectionName(collectionName string) string {
-	invalidatedCollectionName := collectionName + invalidatedEventsCollectionSuffix
-	return invalidatedCollectionName
-}
-
-func storeRootEvent(collectionName string, ctx context.Context, event *models.KeptnContextExtendedCE) error {
-	if collectionName == eventsCollectionName {
-		logger.Debug("Will not store root event because no project has been set in the event")
-		return nil
-	}
-
-	rootEventsForProjectCollection := client.Database(mongoDBName).Collection(collectionName + rootEventCollectionSuffix)
-
-	for _, indexName := range rootEventsIndexes {
-		ensureIndexExistsOnCollection(
-			ctx,
-			rootEventsForProjectCollection,
-			indexName,
-		)
-	}
-
-	result := rootEventsForProjectCollection.FindOne(ctx, bson.M{"shkeptncontext": event.Shkeptncontext})
-	if result.Err() != nil && result.Err() == mongo.ErrNoDocuments {
-		err := storeEventInCollection(event, rootEventsForProjectCollection, ctx)
-		if err != nil {
-			err = fmt.Errorf("Failed to store root event for KeptnContext "+event.Shkeptncontext+": %v", err.Error())
-			return err
-		}
-		logger.Debug("Stored root event for KeptnContext: " + event.Shkeptncontext)
-	} else if result.Err() != nil {
-		// found an already stored root event => check if incoming event precedes the already existing event
-		// if yes, then the new event will be the new root event for this context
-		existingEvent := &models.KeptnContextExtendedCE{}
-
-		err := result.Decode(existingEvent)
-		if err != nil {
-			logger.Error("Could not decode existing root event: " + err.Error())
-			return err
-		}
-
-		if time.Time(existingEvent.Time).After(time.Time(event.Time)) {
-			logger.Debug("Replacing root event for KeptnContext: " + event.Shkeptncontext)
-			_, err := rootEventsForProjectCollection.DeleteOne(ctx, bson.M{"_id": existingEvent.ID})
-			if err != nil {
-				logger.Error("Could not delete previous root event: " + err.Error())
-				return err
-			}
-			err = storeEventInCollection(event, rootEventsForProjectCollection, ctx)
-			if err != nil {
-				err = fmt.Errorf("Failed to store root event for KeptnContext "+event.Shkeptncontext+": %v", err.Error())
-				return err
-			}
-			logger.Debug("Stored new root event for KeptnContext: " + event.Shkeptncontext)
-		}
-	}
-
-	logger.Info("Root event for KeptnContext " + event.Shkeptncontext + " already exists in collection")
-	return nil
-}
-
-func storeEventInCollection(event *models.KeptnContextExtendedCE, collection *mongo.Collection, ctx context.Context) error {
-	eventInterface, err := transformEventToInterface(event)
-	if err != nil {
-		return err
-	}
-
-	_, err = collection.InsertOne(ctx, eventInterface)
-	if err != nil {
-		err := fmt.Errorf("Failed to store root event for KeptnContext "+event.Shkeptncontext+": %v", err.Error())
-		return err
-	}
-
-	return nil
-}
-
-func storeContextToProjectMapping(event *models.KeptnContextExtendedCE, ctx context.Context, collectionName string) error {
-	if collectionName == eventsCollectionName {
-		logger.Debug("Will not store mapping between context and project because no project has been set in the event")
-		return nil
-	}
-
-	logger.Debug(fmt.Sprintf("Storing mapping %s->%s", event.Shkeptncontext, collectionName))
-	contextToProjectCollection := client.Database(mongoDBName).Collection(contextToProjectCollection)
-
-	_, err := contextToProjectCollection.InsertOne(ctx,
-		bson.M{"_id": event.Shkeptncontext, "shkeptncontext": event.Shkeptncontext, "project": collectionName},
-	)
-	if err != nil {
-		if writeErr, ok := err.(mongo.WriteException); ok {
-			if len(writeErr.WriteErrors) > 0 && writeErr.WriteErrors[0].Code == 11000 { // 11000 = duplicate key error
-				logger.Info("Mapping " + event.Shkeptncontext + "->" + collectionName + " already exists in collection")
-			}
-		} else {
-			err := fmt.Errorf("Failed to store mapping "+event.Shkeptncontext+"->"+collectionName+": %v", err.Error())
-			logger.Error(err.Error())
-			return err
-		}
-	}
-
-	return nil
-}
-
-func dropProjectEvents(event *models.KeptnContextExtendedCE) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	projectName := getProjectOfEvent(event)
-	projectCollection := client.Database(mongoDBName).Collection(projectName)
-	rootEventsCollection := client.Database(mongoDBName).Collection(projectName + rootEventCollectionSuffix)
-	invalidatedEventsCollection := client.Database(mongoDBName).Collection(getInvalidatedCollectionName(projectName))
-
-	for _, indexName := range projectEventsIndexes {
-		skipCreateIndex[getIndexIDForCollection(projectCollection.Name(), indexName)] = false
-	}
-	for _, indexName := range rootEventsIndexes {
-		skipCreateIndex[getIndexIDForCollection(rootEventsCollection.Name(), indexName)] = false
-	}
-	for _, indexName := range invalidatedEventsIndexes {
-		skipCreateIndex[getIndexIDForCollection(invalidatedEventsCollection.Name(), indexName)] = false
-	}
-
-	logger.Debug(fmt.Sprintf("Delete all events of project %s", projectName))
-
-	err := projectCollection.Drop(ctx)
-	if err != nil {
-		err := fmt.Errorf("failed to drop collection %s: %v", projectCollection.Name(), err)
-		logger.Error(err.Error())
-		return err
-	}
-
-	logger.Debug(fmt.Sprintf("Delete all root events of project %s", projectName))
-
-	err = rootEventsCollection.Drop(ctx)
-	if err != nil {
-		err := fmt.Errorf("failed to drop collection %s: %v", rootEventsCollection.Name(), err)
-		logger.Error(err.Error())
-		return err
-	}
-
-	logger.Debug(fmt.Sprintf("Delete all invalidated events of project %s", projectName))
-
-	skipCreateIndex[invalidatedEventsCollection.Name()+"-triggeredid"] = false
-	err = invalidatedEventsCollection.Drop(ctx)
-	if err != nil {
-		// log the error but continue
-		err := fmt.Errorf("failed to drop collection %s: %v", invalidatedEventsCollection.Name(), err)
-		logger.Error(err.Error())
-	}
-
-	logger.Debug(fmt.Sprintf("Delete context-to-project mappings of project %s", projectName))
-	contextToProjectCollection := client.Database(mongoDBName).Collection(contextToProjectCollection)
-	if _, err := contextToProjectCollection.DeleteMany(ctx, bson.M{"project": projectName}); err != nil {
-		err := fmt.Errorf("failed to delete context-to-project mapping for project %s: %v", projectName, err)
-		logger.Error(err.Error())
-		return err
-	}
-
-	return nil
-}
-
-func transformEventToInterface(event interface{}) (interface{}, error) {
-	data, err := json.Marshal(event)
-	if err != nil {
-		err := fmt.Errorf("failed to marshal event: %v", err)
-		return nil, err
-	}
-
-	var eventInterface interface{}
-	err = json.Unmarshal(data, &eventInterface)
-	if err != nil {
-		err := fmt.Errorf("failed to unmarshal event: %v", err)
-		return nil, err
-	}
-
-	return eventInterface, nil
-}
-
-func getProjectOfEvent(event *models.KeptnContextExtendedCE) string {
-	collectionName := eventsCollectionName
-	// check if the data object contains the project name.
-	// if yes, store the event in the collection for the project, otherwise in /events
-	eventData, ok := event.Data.(map[string]interface{})
-	if ok && eventData["project"] != nil {
-		collectionNameStr, ok := eventData["project"].(string)
-		if ok && collectionNameStr != "" {
-			collectionName = collectionNameStr
-		}
-	}
-
-	return collectionName
-}
-
-// GetEvents returns all events from the data store sorted by time
-func GetEvents(params event.GetEventsParams) (*event.GetEventsOKBody, error) {
-	logger.Debug("getting events from the data store")
-
-	if err := ensureDBConnection(); err != nil {
-		err := fmt.Errorf("failed to establish MongoDB connection: %v", err)
-		logger.Error(err.Error())
-		return nil, err
-	}
-
-	searchOptions := getSearchOptions(params)
-
-	onlyRootEvents := params.Root != nil
-	collectionName, err := getCollectionNameForQuery(searchOptions)
-	if err != nil {
-		return nil, err
-	} else if collectionName == "" {
-		return &event.GetEventsOKBody{
-			Events:      nil,
-			NextPageKey: "0",
-			PageSize:    0,
-			TotalCount:  0,
-		}, nil
-	}
-	result, err := findInDB(collectionName, *params.PageSize, params.NextPageKey, onlyRootEvents, searchOptions)
+func (erh *EventRequestHandler) GetEventsByType(params event.GetEventsByTypeParams) (*event.GetEventsByTypeOKBody, error) {
+	events, err := erh.eventRepo.GetEventsByType(params)
 	if err != nil {
 		return nil, err
 	}
-
-	return (*event.GetEventsOKBody)(result), nil
-}
-
-type getEventsResult struct {
-	// Events
-	Events []*models.KeptnContextExtendedCE `json:"events"`
-	// Pointer to the next page
-	NextPageKey string `json:"nextPageKey,omitempty"`
-	// Size of the returned page
-	PageSize int64 `json:"pageSize,omitempty"`
-	// Total number of events
-	TotalCount int64 `json:"totalCount,omitempty"`
-}
-
-func aggregateFromDB(collectionName string, pipeline mongo.Pipeline) (*getEventsResult, error) {
-	collection := client.Database(mongoDBName).Collection(collectionName)
-
-	result := &getEventsResult{
-		Events: []*models.KeptnContextExtendedCE{},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	cur, err := collection.Aggregate(ctx, pipeline)
-
-	if err != nil {
-		logger.WithError(err).Error("error finding elements in events collection")
-		return nil, err
-	}
-	// close the cursor after the function has completed to avoid memory leaks
-	defer func() {
-		if err := cur.Close(ctx); err != nil {
-			logger.WithError(err).Error("could not close cursor")
-		}
-	}()
-	result.Events = formatEventResults(ctx, cur)
-
-	return result, nil
-}
-
-func findInDB(collectionName string, pageSize int64, nextPageKeyStr *string, onlyRootEvents bool, searchOptions bson.M) (*getEventsResult, error) {
-	var newNextPageKey int64
-	var nextPageKey int64 = 0
-	if nextPageKeyStr != nil {
-		tmpNextPageKey, _ := strconv.Atoi(*nextPageKeyStr)
-		nextPageKey = int64(tmpNextPageKey)
-		newNextPageKey = nextPageKey + pageSize
-	} else {
-		newNextPageKey = pageSize
-	}
-
-	var sortOptions *options.FindOptions
-
-	if onlyRootEvents {
-		collectionName = collectionName + rootEventCollectionSuffix
-	}
-	if pageSize > 0 {
-		sortOptions = options.Find().SetSort(bson.D{{Key: "time", Value: -1}}).SetSkip(nextPageKey).SetLimit(pageSize)
-	} else {
-		sortOptions = options.Find().SetSort(bson.D{{Key: "time", Value: -1}})
-	}
-
-	collection := client.Database(mongoDBName).Collection(collectionName)
-
-	var result getEventsResult
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	totalCount, err := collection.CountDocuments(ctx, searchOptions)
-	if err != nil {
-		logger.WithError(err).Error("error counting elements in events collection")
-	}
-
-	cur, err := collection.Find(ctx, searchOptions, sortOptions)
-
-	if err != nil {
-		logger.WithError(err).Error("error finding elements in events collection")
-		return nil, err
-	}
-	// close the cursor after the function has completed to avoid memory leaks
-	defer func() {
-		if err := cur.Close(ctx); err != nil {
-			logger.WithError(err).Error("could not close cursor")
-		}
-	}()
-	result.Events = formatEventResults(ctx, cur)
-
-	result.PageSize = pageSize
-	result.TotalCount = totalCount
-
-	if newNextPageKey < totalCount {
-		result.NextPageKey = strconv.FormatInt(newNextPageKey, 10)
-	}
-
-	return &result, nil
-}
-
-func formatEventResults(ctx context.Context, cur *mongo.Cursor) []*models.KeptnContextExtendedCE {
-	events := []*models.KeptnContextExtendedCE{}
-	for cur.Next(ctx) {
-		var outputEvent interface{}
-		err := cur.Decode(&outputEvent)
-		if err != nil {
-			logger.WithError(err).Error("failed to decode event")
-			continue
-		}
-		outputEvent, err = flattenRecursively(outputEvent)
-		if err != nil {
-			logger.WithError(err).Error("failed to flatten")
-			continue
-		}
-
-		data, _ := json.Marshal(outputEvent)
-
-		var keptnEvent models.KeptnContextExtendedCE
-		err = keptnEvent.UnmarshalJSON(data)
-		if err != nil {
-			logger.WithError(err).Error("failed to unmarshal")
-			continue
-		}
-
-		// backwards compatibility: transform evaluation-done events to evaluation.finished events
-		if keptnEvent.Type == keptn07EvaluationDoneEventType {
-			if err := transformEvaluationDonEvent(&keptnEvent); err != nil {
-				logger.Error(err.Error())
-				continue
-			}
-		}
-
-		events = append(events, &keptnEvent)
-	}
-	return events
-}
-
-func transformEvaluationDonEvent(keptnEvent *models.KeptnContextExtendedCE) error {
-	eventMap := map[string]interface{}{}
-	convertedEvent := &keptnv2.EvaluationFinishedEventData{}
-	if err := keptnv2.Decode(keptnEvent.Data, &eventMap); err != nil {
-		return fmt.Errorf("failed to transform evaluation-done event to evaluation.finished event %v", err)
-	}
-	if err := keptnv2.Decode(keptnEvent.Data, convertedEvent); err != nil {
-		return fmt.Errorf("failed to transform evaluation-done event to evaluation.finished event %v", err)
-	}
-	if eventMap["evaluationdetails"] != nil {
-		evaluationDetails := &keptnv2.EvaluationDetails{}
-		if err := keptnv2.Decode(eventMap["evaluationdetails"], evaluationDetails); err != nil {
-			return fmt.Errorf("failed to transform evaluationDetails of evaluation-done event: %v", err)
-		}
-		convertedEvent.Evaluation = *evaluationDetails
-	}
-	keptnEvent.Type = models.Type(keptnv2.GetFinishedEventType(keptnv2.EvaluationTaskName))
-	keptnEvent.Data = convertedEvent
-	keptnEvent.Specversion = "1.0"
-	return nil
-}
-
-func getCollectionNameForQuery(searchOptions bson.M) (string, error) {
-	collectionName := eventsCollectionName
-	if searchOptions["data.project"] != nil && searchOptions["data.project"] != "" {
-		// if a project has been specified, query the collection for that project
-		collectionName = searchOptions["data.project"].(string)
-	} else if searchOptions["shkeptncontext"] != nil && searchOptions["shkeptncontext"] != "" {
-		var err error
-		collectionName, err = getProjectForContext(searchOptions["shkeptncontext"].(string))
-		if err != nil {
-			if err == mongo.ErrNoDocuments {
-				logger.Info("no project found for shkeptncontext")
-				return unmappedEventsCollectionName, nil
-			}
-			logger.WithError(err).Error("error loading project for shkeptncontext")
-			return "", err
-		}
-	}
-
-	return collectionName, nil
-}
-
-func getProjectForContext(keptnContext string) (string, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	contextToProjectCollection := client.Database(mongoDBName).Collection(contextToProjectCollection)
-	result := contextToProjectCollection.FindOne(ctx, bson.M{"shkeptncontext": keptnContext})
-	var resultMap bson.M
-	err := result.Decode(&resultMap)
-	if err != nil {
-		return "", err
-	}
-	if project, ok := resultMap["project"].(string); !ok {
-		return "", errors.New("Could not find entry for keptnContext: " + keptnContext)
-	} else {
-		return project, nil
-	}
-}
-
-func getSearchOptions(params event.GetEventsParams) bson.M {
-	searchOptions := bson.M{}
-	if params.KeptnContext != nil {
-		searchOptions["shkeptncontext"] = *params.KeptnContext
-	}
-	if params.Type != nil {
-		// for backwards compatibility: if evaluation.finished events are queried, also retrieve evaluation-done events
-		searchOptions = setEventTypeMatchCriteria(*params.Type, searchOptions)
-	}
-	if params.Source != nil {
-		searchOptions["source"] = *params.Source
-	}
-	if params.Project != nil {
-		searchOptions["data.project"] = *params.Project
-	}
-	if params.Stage != nil {
-		searchOptions["data.stage"] = *params.Stage
-	}
-	if params.Service != nil {
-		searchOptions["data.service"] = *params.Service
-	}
-	if params.EventID != nil {
-		searchOptions["id"] = *params.EventID
-	}
-	if params.FromTime != nil {
-		if params.BeforeTime == nil {
-			searchOptions["time"] = bson.M{
-				"$gt": *params.FromTime,
-			}
-		} else {
-			searchOptions["$and"] = []bson.M{
-				{"time": bson.M{"$gt": *params.FromTime}},
-				{"time": bson.M{"$lt": *params.BeforeTime}},
-			}
-		}
-	}
-	if params.BeforeTime != nil && params.FromTime == nil {
-		searchOptions["time"] = bson.M{
-			"$lt": *params.BeforeTime,
-		}
-	}
-
-	return searchOptions
-}
-
-func setEventTypeMatchCriteria(eventType string, searchOptions bson.M) bson.M {
-	if eventType == keptnv2.GetFinishedEventType(keptnv2.EvaluationTaskName) {
-		searchOptions["$or"] = []bson.M{
-			{"type": eventType},
-			{"type": keptn07EvaluationDoneEventType},
-		}
-	} else {
-		searchOptions["type"] = eventType
-	}
-	return searchOptions
-}
-
-func flattenRecursively(i interface{}) (interface{}, error) {
-	if _, ok := i.(bson.D); ok {
-		d := i.(bson.D)
-		myMap := d.Map()
-		flat, err := flatten.Flatten(myMap, "", flatten.RailsStyle)
-		if err != nil {
-			logger.WithError(err).Error("could not flatten element")
-			return nil, err
-		}
-		for k, v := range flat {
-			res, err := flattenRecursively(v)
-			if err != nil {
-				return nil, err
-			}
-			if k == "eventContext" {
-				flat[k] = nil
-			} else {
-				flat[k] = res
-			}
-		}
-		return flat, nil
-	} else if _, ok := i.(bson.A); ok {
-		a := i.(bson.A)
-		for i := 0; i < len(a); i++ {
-			res, err := flattenRecursively(a[i])
-			if err != nil {
-				return nil, err
-			}
-			a[i] = res
-		}
-		return a, nil
-	}
-
-	return i, nil
-}
-
-// MinimumFilterNotProvided indicates that the minimum requirements for the filter have not been met
-var MinimumFilterNotProvided = errors.New("must provide a filter containing at least one of the following properties: 'data.project' or 'shkeptncontext'")
-
-// GetEventsByTypeHandlerFunc gets events by their type
-func GetEventsByType(params event.GetEventsByTypeParams) (*event.GetEventsByTypeOKBody, error) {
-	logger.Debug(fmt.Sprintf("getting %s events from the data store", params.EventType))
-
-	if err := ensureDBConnection(); err != nil {
-		err := fmt.Errorf("failed to establish MongoDB connection: %v", err)
-		logger.Error(err.Error())
-		return nil, err
-	}
-	// TODO: this will be moved
-
-	if params.Filter == nil {
-		return nil, MinimumFilterNotProvided
-	}
-
-	matchFields := parseFilter(*params.Filter)
-	if !validateFilter(matchFields) {
-		return nil, MinimumFilterNotProvided
-	}
-
-	matchFields = setEventTypeMatchCriteria(params.EventType, matchFields)
-
-	if params.FromTime != nil {
-		matchFields["time"] = bson.M{
-			"$gt": *params.FromTime,
-		}
-	}
-
-	collectionName, err := getCollectionNameForQuery(matchFields)
-	if err != nil {
-		return nil, err
-	} else if collectionName == "" {
-		return &event.GetEventsByTypeOKBody{
-			Events: nil,
-		}, nil
-	}
-
-	var allEvents *getEventsResult
-
-	if params.ExcludeInvalidated != nil && *params.ExcludeInvalidated {
-		aggregationPipeline := getAggregationPipeline(params, collectionName, matchFields)
-		allEvents, err = aggregateFromDB(collectionName, aggregationPipeline)
-	} else {
-		allEvents, err = findInDB(collectionName, *params.Limit, nil, false, matchFields)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return &event.GetEventsByTypeOKBody{Events: allEvents.Events}, nil
-}
-
-func getAggregationPipeline(params event.GetEventsByTypeParams, collectionName string, matchFields bson.M) mongo.Pipeline {
-
-	matchStage := bson.D{
-		{Key: "$match", Value: matchFields},
-	}
-
-	lookupStage := bson.D{
-		{Key: "$lookup", Value: bson.M{
-			"from": getInvalidatedCollectionName(collectionName),
-			"let": bson.M{
-				"event_id":          "$id",
-				"event_triggeredid": "$triggeredid",
-			},
-			"pipeline": []bson.M{
-				{
-					"$match": bson.M{
-						"$expr": bson.M{
-							"$or": []bson.M{
-								{
-									// backwards-compatibility to 0.7.x -> triggeredid of .invalidated event refers to the id of the evaluation-done event
-									"$eq": []string{"$triggeredid", "$$event_id"},
-								},
-								{
-									// logic for 0.8: triggeredid of .invalidated event refers to the triggeredid of the evaluation.finished event (both are related to the same .triggered event)
-									"$eq": []string{"$triggeredid", "$$event_triggeredid"},
-								},
-							},
-						},
-					},
-				},
-				{
-					"$limit": 1,
-				},
-			},
-			"as": "invalidated",
-		}},
-	}
-
-	matchInvalidatedStage := bson.D{
-		{Key: "$match", Value: bson.M{
-			"invalidated": bson.M{
-				"$size": 0,
-			},
-		}},
-	}
-	sortStage := bson.D{
-		{Key: "$sort", Value: bson.M{
-			"time": -1,
-		}},
-	}
-	var aggregationPipeline mongo.Pipeline
-	if params.Limit != nil && *params.Limit > 0 {
-		limitStage := bson.D{
-			{Key: "$limit", Value: *params.Limit},
-		}
-		aggregationPipeline = mongo.Pipeline{matchStage, lookupStage, matchInvalidatedStage, sortStage, limitStage}
-	} else {
-		aggregationPipeline = mongo.Pipeline{matchStage, lookupStage, matchInvalidatedStage, sortStage}
-	}
-
-	return aggregationPipeline
-}
-
-func getInvalidatedEventType(eventType string) string {
-	var invalidatedEventType string
-
-	split := strings.Split(eventType, ".")
-	invalidatedEventType = split[0]
-	for i := 1; i < len(split)-1; i = i + 1 {
-		invalidatedEventType = invalidatedEventType + "." + split[i]
-	}
-
-	invalidatedEventType = invalidatedEventType + ".invalidated"
-	return invalidatedEventType
-}
-
-func validateFilter(searchOptions bson.M) bool {
-	if (searchOptions["data.project"] == nil || searchOptions["data.project"] == "") && (searchOptions["shkeptncontext"] == nil || searchOptions["shkeptncontext"] == "") {
-		return false
-	}
-
-	return true
-}
-
-func parseFilter(filter string) bson.M {
-	filterObject := bson.M{}
-	keyValues := strings.Split(filter, " AND ")
-
-	for _, keyValuePair := range keyValues {
-		split := strings.Split(keyValuePair, ":")
-		if len(split) == 2 {
-			splitValue := strings.Split(split[1], ",")
-			if len(splitValue) == 1 {
-				filterObject[split[0]] = split[1]
-			} else {
-				filterObject[split[0]] = bson.M{
-					"$in": splitValue,
-				}
-			}
-		}
-	}
-
-	return filterObject
+	return &event.GetEventsByTypeOKBody{Events: events.Events}, nil
 }
